Done with HTML element creation tracking.
Added Utils/config.js for managing the configuration.
Added Utils/logger.js for managing the logging process.
Added Tracker/tracker.js that contains the Tracker class which is used
for tracking the HTML element and for connecting the Tainter and
DebuggerWrapper classes.
Reworked DebuggerWrapper class with custom handlers of onEnterFrame and
onStep events.
Now the DebuggerWrapper and the Tracker classes are able to keep track
of the creation of the needed HTML element, and also to change its name.