Added some more test suites
Added “literals” test suite to test the tainting of literal values.
Added “object_operation” test suite to test the tainting and
propagation of objects and their properties.
Rewritten the proxy objects in such way that now they can keep track of
tainted and untainted values. This is done by using a set of untainted
values, that gets modified whenever a setter of one of the object’s
properties is called.