Changed the script's output
Reworked the output that the script creates.
Replaced the way in which the tracker understands that all webpage code
was executed. Previously we added a debugger statement at the end of
the page and outputted it the onDebuggerStatement handler. Now we are
using the second parameter of the slimer’s webpage.open() function.