Added tainting for member functions
Now the results of calls of member functions of tainted objects
are also tainted. This effect is achieved by adding a replacer
for the x.f(args) expression (which now gets rewritten into smth.
like __call__(x, 'f', args)). This workaround is needed because of
some quircks with the this keyword passing to functions.